requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==1.10.13
pymongo[srv]==4.6.1
python-dotenv==1.0.1
streamlit==1.31.0
pandas==2.2.0
